Add Clash.Normalize.Transformations.EtaExpand
All transformations concerning eta expansion now live in a separate
module. This is part of work to make the normalization modules
quicker to compile.
